code cleanup1
