+++ CloudShellApi/QS.ALM.CloudShellApi.csproj
@@ -65,6 +65,7 @@
     <Compile Include="Api.cs" />
     <Compile Include="APIContracts.cs" />
     <Compile Include="Config.cs" />
+    <Compile Include="RunStatusManager.cs" />
     <Compile Include="SettingsFile.cs" />
     <Compile Include="Constants.cs" />
     <Compile Include="Logger.cs" />
